Make recording service just record straight to WAV.
This means that the contents don't have to be copied to a whole new
file, like what the previous implementation did - this means that we
are free to save large files (the WAV format allows up to about 4GB).
It now creates a WAV as soon as recording starts, with the necessary
headers (the two size headers are added on when recording stops).
"Saving" it just moves the WAV file to the appropriate directory (with
the appropriate filename).

This was largely helped by
https://gist.github.com/kmark/d8b1b01fb0d2febf5770