Add launchAudioImplicitIntent function to util.File, and use it
When tapping on past recording, it will launch an implicit intent.
Note that when testing on my device, this opens it successfully in VLC
but not Podcast Addict, Google Play Music, or YouTube Music.
However, this is the same when opening a WAV file from my file explorer,
so I think it is ok.