Rearrange schedule recording validation
Move nameExists below !startsInTheFuture(), so that if it doesn't start
in the future and the name already exists, it shows the error about
starting in the future. Previously, it would show the name exists error,
which the user was able to schedule the recording from.
This lead to it being possible to schedule recordings in the past.