Only save the user-specified buffer.
If the user specifies a time that is the max (actual start or end), then
use seconds for the actual start or end. Otherwise, just use 0s for
seconds (and milliseconds).

To do this, a cropped file is created in internal storage, headers
written, channels are used to copy over the relevant bytes, then size
headers updated. This cropped file is then moved (or copied)

- Remove filename field if user isn't using custom filenames
- Move writeHeaders and updateHeaders into util